rename: 패키지 구조 변경 - #5
- 도메인 별 패키지 구조 적용
