+++ src/main/java/org/mju_likelion/festival/admin/domain/AdminRole.java
@@ -1,4 +1,4 @@
-package org.mju_likelion.festival.domain.admin;
+package org.mju_likelion.festival.admin.domain;
 
 public enum AdminRole {
   STUDENT_COUNCIL, BOOTH_MANAGER
