+++ src/main/java/org/mju_likelion/festival/image/domain/Image.java
@@ -1,6 +1,6 @@
-package org.mju_likelion.festival.domain.image;
+package org.mju_likelion.festival.image.domain;
 
-import static org.mju_likelion.festival.domain.common.constant.ColumnLengths.IMAGE_URL_LENGTH;
+import static org.mju_likelion.festival.common.domain.constant.ColumnLengths.IMAGE_URL_LENGTH;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -12,10 +12,10 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.mju_likelion.festival.domain.announcement.AnnouncementImage;
-import org.mju_likelion.festival.domain.booth.BoothImage;
-import org.mju_likelion.festival.domain.common.BaseEntity;
-import org.mju_likelion.festival.domain.lost_item.LostItemImage;
+import org.mju_likelion.festival.announcement.domain.AnnouncementImage;
+import org.mju_likelion.festival.booth.domain.BoothImage;
+import org.mju_likelion.festival.common.domain.BaseEntity;
+import org.mju_likelion.festival.lost_item.domain.LostItemImage;
 
 @Getter
 @Builder
