+++ src/main/java/org/mju_likelion/festival/lost_item/domain/LostItemImage.java
@@ -1,4 +1,4 @@
-package org.mju_likelion.festival.domain.term;
+package org.mju_likelion.festival.lost_item.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,21 +8,21 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.mju_likelion.festival.domain.common.BaseEntity;
-import org.mju_likelion.festival.domain.user.User;
+import org.mju_likelion.festival.common.domain.BaseEntity;
+import org.mju_likelion.festival.image.domain.Image;
 
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Entity(name = "term_user")
-public class TermUser extends BaseEntity {
+@Entity(name = "lost_item_image")
+public class LostItemImage extends BaseEntity {
 
   @ManyToOne(optional = false, fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id", nullable = false)
-  private User user;
+  @JoinColumn(name = "lost_item_id", nullable = false)
+  private LostItem lostItem;
 
   @ManyToOne(optional = false, fetch = FetchType.LAZY)
-  @JoinColumn(name = "term_id", nullable = false)
-  private Term term;
+  @JoinColumn(name = "image_id", nullable = false)
+  private Image image;
 }
